requirements.txt
matplotlib
numpy
openslide-python
pillow
pyyaml
scikit-image
scikit-plot
sklearn
torch
torchvision
tqdm
